requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
librosa
numpy
torch
transformers
openai-whisper
soundfile
